+++ LootViewer.lua
@@ -1,7 +1,7 @@
 local addonName, LV = ...
 
 LV.name = addonName
-LV.version = "12.1.5"
+LV.version = "12.1.6-2"
 LV.frame = CreateFrame("Frame")
 LV.eventHandlers = {}
 LV.modules = {}